Fixed findBooksByCategoryId
