+++ src/main/java/com/foalex/bookstore/service/impl/CategoryServiceImpl.java
@@ -1,6 +1,5 @@
 package com.foalex.bookstore.service.impl;
 
-import com.foalex.bookstore.dto.book.BookDtoWithoutCategories;
 import com.foalex.bookstore.dto.category.CategoryDto;
 import com.foalex.bookstore.dto.category.CreateCategoryRequestDto;
 import com.foalex.bookstore.dto.category.UpdateCategoryRequestDto;
@@ -39,43 +38,26 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public CategoryDto getById(Long id) {
-        return categoryRepository.findById(id)
-                .map(categoryMapper::toDto)
-                .orElseThrow(() -> new EntityNotFoundException("""
-                             Get operation failed. 
-                             Category with id %d doesn't exist.""".formatted(id)));
+        return categoryMapper.toDto(categoryByOperationAndId("Get", id));
     }
 
     @Override
     public CategoryDto update(Long id, UpdateCategoryRequestDto requestDto) {
-        Category category = categoryRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("""
-                             Update operation failed. 
-                             Category with id %d doesn't exist.""".formatted(id)));
+        Category category = categoryByOperationAndId("Update", id);
         categoryMapper.mapUpdateRequestToCategory(requestDto, category);
         return categoryMapper.toDto(categoryRepository.save(category));
     }
 
     @Override
     public void delete(Long id) {
-        Category category = categoryRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("""
-                             Delete operation failed. 
-                             Category with id %d doesn't exist.""".formatted(id)));
-        categoryRepository.delete(category);
+        categoryRepository.delete(categoryByOperationAndId("Delete", id));
     }
 
-    @Override
-    public List<BookDtoWithoutCategories> findBooksByCategoryId(Long id, Pageable pageable) {
-        if (!categoryRepository.existsById(id)) {
-            return bookRepository.findAllByCategoriesId(id, pageable)
-                    .stream()
-                    .map(bookMapper::toDtoWithoutCategories)
-                    .toList();
-        } else {
-            throw new EntityNotFoundException("""
-                             Find Books by Category operation failed. 
-                             Category with id %d doesn't exist.""".formatted(id));
-        }
+    private Category categoryByOperationAndId(String operation, Long id) {
+        return categoryRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("""
+                             %s operation failed.
+                             Category with id %d doesn't exist."""
+                        .formatted(operation, id)));
     }
 }
